+++ CHANGES.md
@@ -1,5 +1,9 @@
 # CHANGES #
 
+### 2023-04-02
+
+* log all urls containing keywords with the keywords listed: url,keyword1,keyword2,...
+
 ### 2023-03-27
 
 * generate "products.txt" out of "products.csv" directly